fix: update path and naming folder
