+++ Messenger/App.tsx
@@ -5,11 +5,17 @@ import { Provider } from "react-redux";
 import store from "./ReducersAndActions/ConfigureStore/ConfigureStore";
 import { SafeAreaProvider } from "react-native-safe-area-context";
 import Navigation from "./Navigation/Navigation";
-import { View } from "react-native";
 
 export default function App() {
   StatusBar.setBarStyle("dark-content");
-  return (<View></View>);
+  return (
+    <SafeAreaProvider>
+      <StatusBar translucent backgroundColor="transparent" />
+      <Provider store={store}>
+        <Navigation />
+      </Provider>
+    </SafeAreaProvider>
+  );
 }
 registerRootComponent(App);
 AppRegistry.registerComponent("10", () => App);
