+++ Messenger/Initialization/initializationMessage.ts
@@ -6,9 +6,12 @@ import MainChat from "../dao/Models/Chats/MainChat";
 import { getRandomElementsFromArray, getRandomNumber } from './functions';
 
 
-export function addMessages(chat: Chat, count: number, users: User[], texts: string[] = []) {
+export function addMessages(chat: Chat, count: number, users: User[] = [], texts: string[] = []) {
     let idMessageToCreate = 0
+    if(chat.messages == undefined) chat.messages = []
     if (chat.messages.length > 0) idMessageToCreate = chat.messages[chat.messages.length - 1].messageId! + 1;
+    
+    if(users.length == undefined) users = []
 
     if (users.length === 0) throw new Error("must be more than 1 users");
 
@@ -20,7 +23,7 @@ export function addMessages(chat: Chat, count: number, users: User[], texts: str
         else
             content = texts[getRandomNumber(texts.length)];
 
-        const message = new Message(users[getRandomNumber(users.length)], content, new Date(), EMessageType.text);
+        const message = new Message(users[getRandomNumber(users.length)], content, EMessageType.text);
         message.messageId = idMessageToCreate++;
         // Additional properties can be set if needed
         if (Math.random() < 0.15) message.messageResponseId = idMessageToCreate - 2; // Set response ID to the previous message ID
@@ -32,27 +35,27 @@ export function addMessages(chat: Chat, count: number, users: User[], texts: str
 export function initializationLastWatchedMessageChat(chat: MainChat) {
     if (chat.branches.length > 0) {
         for (let branch of chat.branches) {
-            initializationLastWatchedMessageBranch(branch, chat.users);
+            //initializationLastWatchedMessageBranch(branch, chat.users);
         }
     } else {
         for (let user of chat.users) {
-            chat.lastWatchedMessage.push({ user: user, value: chat.messages[getRandomNumber(chat.messages.length)] })
+            //chat.lastWatchedMessage.push({ user: user, value: chat.messages[getRandomNumber(chat.messages.length)] })
         }
     }
 
 }
-export function initializationLastWatchedMessageBranch(chat: Chat, users: User[]) {
-    if (chat.branches.length > 0) {
-        for (let branch of chat.branches) {
-            initializationLastWatchedMessageBranch(branch, users)
-        }
-    } else {
-        for (let user of users) {
-            chat.lastWatchedMessage.push({ user: user, value: chat.messages[getRandomNumber(chat.messages.length)] })
-        }
-    }
-
-}
+//export function initializationLastWatchedMessageBranch(chat: Chat, users: User[]) {
+//    if (chat.branches.length > 0) {
+//        for (let branch of chat.branches) {
+//            initializationLastWatchedMessageBranch(branch, users)
+//        }
+//    } else {
+//        for (let user of users) {
+//            //chat.lastWatchedMessage.push({ user: user, value: chat.messages[getRandomNumber(chat.messages.length)] })
+//        }
+//    }
+//
+//}
 export function initializationPinnedMessage(chat: Chat, count_max?: number) {
     chat.pinnedMessage.push(...getRandomElementsFromArray(chat.messages, count_max));
 }
