+++ Messenger/dao/Models/Chats/IMessageLog.ts
